move ExeWrapper into its own file
